Fixed some things that prevented the program from functioning due to poor checking of directories (assuming they were there)

+++ PDB2MC/minecraft_functions.py
@@ -108,13 +108,16 @@ def delete_nbt_mcfunctions_from_dir(save_directory):
 
 
 def delete_mcfunctions(directory, name):
-    for filename in os.listdir(directory):
-        if filename.startswith("build_" + name) and filename.endswith(".mcfunction"):
-            os.remove(os.path.join(directory, filename))
-        if filename.startswith(name) and filename.endswith(".mcfunction"):
-            os.remove(os.path.join(directory, filename))
-        if filename.startswith(name) and filename.endswith(".nbt"):
-            os.remove(os.path.join(directory, filename))
+    if os.path.exists(directory):
+        for filename in os.listdir(directory):
+            if filename.startswith("build_" + name) and filename.endswith(".mcfunction"):
+                os.remove(os.path.join(directory, filename))
+            if filename.startswith(name) and filename.endswith(".mcfunction"):
+                os.remove(os.path.join(directory, filename))
+            if filename.startswith(name) and filename.endswith(".nbt"):
+                os.remove(os.path.join(directory, filename))
+    else:
+        print(f"Directory {directory} does not exist.")
 
 
 def delete_old_files(mc_dir, pdb_name):
@@ -126,6 +129,7 @@ def delete_old_files(mc_dir, pdb_name):
     # In the parallel directory named "function", delete the old mcfunction
     base_dir = os.path.dirname(mc_dir)  # Get the directory containing the "functions" directory
     new_dir = os.path.join(base_dir, "function")  # Path for the new "function" directory
+    new_dir = os.path.normpath(new_dir)
     delete_mcfunctions(new_dir, "z" + pdb_name.lower())
     delete_mcfunctions(new_dir, "build_" + pdb_name.lower())
     delete_mcfunctions(new_dir, "clear_" + pdb_name.lower())
@@ -189,12 +193,15 @@ def find_mcfunctions(directory, name):
 
 def find_nbt(directory, name):
     file_list = []
-    for filename in os.listdir(directory):
-        if filename.startswith(name) and filename.endswith(".nbt"):
-            parts = filename.split("_")
-            full = filename.split(".")
-            if len(parts) > 1:
-                file_list.append(full[0])
+    if os.path.exists(directory):
+        for filename in os.listdir(directory):
+            if filename.startswith(name) and filename.endswith(".nbt"):
+                parts = filename.split("_")
+                full = filename.split(".")
+                if len(parts) > 1:
+                    file_list.append(full[0])
+    else:
+        print(f"Directory {directory} does not exist.")
     df = pd.DataFrame({"group": file_list})
     return df
 
@@ -321,7 +328,7 @@ def create_structure_from_mcfunction(name, directory, functions):
     pdb_name = name.lower()
     nbtstructure = NBTStructure()
 
-    directory = directory.split("\datapacks/mcPDB/data/protein/functions")[0]
+    directory = directory.split(r"\datapacks/mcPDB/data/protein/functions")[0]
     directory = directory + "/generated/mc/structures/"
 
     # iterate through each line in functions
@@ -341,9 +348,14 @@ def create_structure_from_mcfunction(name, directory, functions):
 
 def create_nbt(df, name, air, dir, blocks):
 
-    dir = dir.split("\datapacks/mcPDB/data/protein/functions")[0]
+    dir = dir.split(r"\datapacks/mcPDB/data/protein/functions")[0]
     dir = dir + "/generated/mc/structures/"
     nbtstructure = NBTStructure()
+    print("trying to make nbt for ", name, " in ", dir)
+
+    #check if dir exists, if not, create it
+    if not os.path.exists(dir):
+        os.makedirs(dir)
 
     block_dict = blocks
     block_type = 'air' if air else 'block'
@@ -420,7 +432,7 @@ def create_individual_nbt_functions(mcfiles, directory):
 
 
 def create_nbt_delete(pdb_name, mc_dir):
-    nbt_dir = mc_dir.split("\datapacks/mcPDB/data/protein/functions")[0]
+    nbt_dir = mc_dir.split(r"\datapacks/mcPDB/data/protein/functions")[0]
     nbt_dir = nbt_dir + "/generated/mc/structures/"
 
     lower_pdb_name = pdb_name.lower()
@@ -529,7 +541,7 @@ def copy_directory_contents(src, dest):
             shutil.copy2(s, d)
 
 def finish_nbts(mc_dir, config_data, pdb_name):
-    nbt_dir = mc_dir.split("\datapacks/mcPDB/data/protein/functions")[0]
+    nbt_dir = mc_dir.split(r"\datapacks/mcPDB/data/protein/functions")[0]
     nbt_dir = nbt_dir + "/generated/mc/structures/"
 
     mcfiles = find_nbt(nbt_dir, pdb_name.lower())
@@ -546,7 +558,7 @@ def finish_nbts(mc_dir, config_data, pdb_name):
     copy_directory_contents(mc_dir, new_dir)
 
 def finish_delete_nbts(mc_dir, pdb_name):
-    nbt_dir = mc_dir.split("\datapacks/mcPDB/data/protein/functions")[0]
+    nbt_dir = mc_dir.split(r"\datapacks/mcPDB/data/protein/functions")[0]
     nbt_dir = nbt_dir + "/generated/mc/structures/"
 
     mcfiles = find_delete_nbt(nbt_dir, pdb_name.lower())
